backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
email-validator
httpx
python-multipart
alembic
google-auth
google-auth-oauthlib
